+++ app/interactors/update_bank_cells.rb
@@ -6,6 +6,6 @@ class UpdateBankCells
   # context.bank
   # context.cells_info
   def call
-    UpdateBankCellsJob.perform_now(bank_id: context.bank.id, cells_info: context.cells_info)
+    UpdateBankCellsJob.perform_later(bank_id: context.bank.id, cells_info: context.cells_info)
   end
 end