+++ config/application.rb
@@ -31,7 +31,7 @@ module ComCon
     end
 
     config.autoload_paths += %W[#{config.root}/app/jobs]
-    config.active_job.queue_adapter = Rails.env.production? ? :sidekiq : :async
+    config.active_job.queue_adapter = :sidekiq
 
     config.generators.system_tests = nil
     config.action_mailer.preview_path = 'spec/mailers/previews'