+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - user_token#create api v2 endpoint
+- equipment uploading
 
 ### Modified
 - line up page
+- ProfScanner addon
 
 ## [1.2.3] - 2020-01-02
 ### Added